Adjust create and delete test for new url schema
RISDEV-136
